moved controller code to service and fixed a bug in userService
