+++ src/main/java/com/forge/revature/controllers/AboutMeController.java
@@ -1,12 +1,9 @@
 package com.forge.revature.controllers;
 
 import java.util.List;
-import java.util.Optional;
-
 import com.forge.revature.models.AboutMe;
-import com.forge.revature.repo.AboutMeRepo;
-
-import org.springframework.beans.factory.annotation.Autowired;
+import com.forge.revature.services.AboutMeService;
+import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -21,29 +18,22 @@ import org.springframework.web.bind.annotation.RestController;
  * @version 1.0
  * 
  * Controller for AboutMe. Has CRUD functionality described per method
+ * Code moved to Service Layer from Controller Layer by Aaron Killeen
  */
 @RestController
 @CrossOrigin(origins = "http://localhost:3000")
 @RequestMapping("api/aboutMe")
+@AllArgsConstructor
 public class AboutMeController {
-    @Autowired
-    AboutMeRepo aboutMeRepo;
-
-    public AboutMeController() {
-    }
-
-    public AboutMeController(AboutMeRepo repo) {
-        this.aboutMeRepo = repo;
-    }
-
+	
+    private AboutMeService aboutMeService;
     /**
      * Retrieves all about mes stored in the database
      * @return List of all about mes in the database in JSON format
      */
     @GetMapping
     public List<AboutMe> getAll() {
-        List<AboutMe> aboutMes = aboutMeRepo.findAll();
-        return aboutMes;
+    	return aboutMeService.getAll();
     }
 
     /**
@@ -53,7 +43,7 @@ public class AboutMeController {
      */
     @PostMapping
     public AboutMe postAboutMe(@RequestBody AboutMe aboutMe) {
-        return aboutMeRepo.save(aboutMe);
+    	return aboutMeService.postAboutMe(aboutMe);
     }
 
     /**
@@ -63,7 +53,7 @@ public class AboutMeController {
      */
     @GetMapping("/{id}")
     public AboutMe getAboutMe(@PathVariable(name = "id") int id) {
-        return aboutMeRepo.findById(id).get();
+    	return aboutMeService.getAboutMe(id);
     }
 
     /**
@@ -73,16 +63,7 @@ public class AboutMeController {
      */
     @PostMapping("/{id}")
     public void updateAboutMe(@RequestBody AboutMe newAboutMe, @PathVariable(name = "id") int aboutMeId) {
-        Optional<AboutMe> oldAboutMe = aboutMeRepo.findById(aboutMeId);
-
-        if (oldAboutMe.isPresent()) {
-
-            oldAboutMe.get().setBio(newAboutMe.getBio());
-            oldAboutMe.get().setEmail(newAboutMe.getEmail());
-            oldAboutMe.get().setPhone(newAboutMe.getPhone());
-
-            aboutMeRepo.save(oldAboutMe.get());
-        }
+    	aboutMeService.updateAboutMe(newAboutMe, aboutMeId);
     }
 
     /**
@@ -91,7 +72,7 @@ public class AboutMeController {
      */
     @DeleteMapping("/{id}")
     public void deleteAboutMe(@PathVariable(name = "id") int aboutMeId) {
-        aboutMeRepo.deleteById(aboutMeId);
+    	aboutMeService.deleteAboutMe(aboutMeId);
     }
 
     /**
@@ -101,12 +82,7 @@ public class AboutMeController {
      */
     @GetMapping("/user/{id}")
     public AboutMe getUserAboutMe(@PathVariable(name = "id") int userId) {
-        Optional<AboutMe> retrievedAboutMe = aboutMeRepo.findByPortfolioUserId(userId);
-
-        if (retrievedAboutMe.isPresent()) {
-            return retrievedAboutMe.get();
-        }
-        return null;
+    	return aboutMeService.getUserAboutMe(userId);
     }
 
     /**
@@ -116,11 +92,6 @@ public class AboutMeController {
      */
     @GetMapping("/portfolio/{id}")
     public AboutMe getPortfolioAboutMe(@PathVariable(name = "id") int portfolioId) {
-        Optional<AboutMe> retrievedAboutMe = aboutMeRepo.findByPortfolioId(portfolioId);
-
-        if (retrievedAboutMe.isPresent()) {
-            return retrievedAboutMe.get();
-        }
-        return null;
+    	return aboutMeService.getPortfolioAboutMe(portfolioId);
     }
 }
