+++ src/main/java/com/forge/revature/repo/UserRepo.java
@@ -4,7 +4,6 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import com.forge.revature.models.User;
@@ -14,4 +13,6 @@ public interface UserRepo extends JpaRepository<User, Integer>{
     public Optional<User> findByEmail(String email);
     
     public List<User> findAllByAdmin(boolean admin);
+    
+    public Optional<User> findById(int id);
 }
